r421: Adding "get_atomicsize" function to block devices. This is used to determine what can be atomically written to disk, but what it actually defines is the maximum change descriptor block size and alignment.

+++ kfs/block_resizer_bd.h
@@ -4,6 +4,6 @@
 #include <inc/types.h>
 #include <kfs/bd.h>
 
-BD_t * block_resizer_bd(BD_t * disk, uint32_t blocksize);
+BD_t * block_resizer_bd(BD_t * disk, uint16_t blocksize);
 
 #endif /* __KUDOS_KFS_BLOCK_RESIZER_BD_H */
